# backend/openapi-swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TodoCamisetas API",
        "description": "API REST para la gestión de camisetas deportivas, clientes B2B, tallas y ventas. Permite administrar el stock de camisetas, gestionar clientes preferenciales y regulares, calcular precios con descuentos y registrar ventas.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/api",
            "description": "Servidor de desarrollo local"
        }
    ],
    "paths": {
...
        "/clientes": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Listar clientes",
                "description": "Obtiene todos los clientes registrados.",
                "operationId": "listarClientes",
                "responses": {
                    "200": {
                        "description": "Listado de clientes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Cliente"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Crear cliente",
                "description": "Registra un nuevo cliente validando el RUT chileno.",
                "operationId": "crearCliente",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ClienteInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cliente creado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Cliente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Datos inválidos o RUT inválido"
                    }
                }
            }
        },
        "/clientes/{cliente}": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Obtener cliente",
                "description": "Obtiene la información de un cliente específico.",
                "operationId": "obtenerCliente",
                "parameters": [
                    {
                        "name": "cliente",
                        "in": "path",
                        "description": "ID del cliente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cliente encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Cliente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Actualizar cliente",
                "description": "Actualiza los datos de un cliente existente.",
                "operationId": "actualizarCliente",
                "parameters": [
                    {
                        "name": "cliente",
                        "in": "path",
                        "description": "ID del cliente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ClienteInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cliente actualizado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Cliente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Datos inválidos o RUT inválido"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Eliminar cliente",
                "description": "Elimina un cliente siempre que no posea ventas asociadas.",
                "operationId": "eliminarCliente",
                "parameters": [
                    {
                        "name": "cliente",
                        "in": "path",
                        "description": "ID del cliente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cliente eliminado correctamente"
                    },
                    "409": {
                        "description": "El cliente posee ventas asociadas"
                    }
                }
            }
        },
        "/clientes/{cliente}/camisetas": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Listar camisetas compradas por un cliente",
                "description": "Obtiene todas las camisetas adquiridas por un cliente.",
                "operationId": "camisetasCliente",
                "parameters": [
                    {
                        "name": "cliente",
                        "in": "path",
                        "description": "ID del cliente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado de camisetas compradas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Camiseta"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Camiseta": {
                "title": "Camiseta",
                "description": "Modelo completo de camiseta deportiva",
                "required": [
                    "titulo",
                    "club",
                    "pais",
                    "tipo",
                    "color",
                    "precio",
                    "stock",
                    "codigo_producto"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "titulo": {
                        "type": "string",
                        "example": "Camiseta Local 2025 - Selección Chilena"
                    },
                    "club": {
                        "type": "string",
                        "example": "Selección Chilena"
                    },
                    "pais": {
                        "type": "string",
                        "example": "Chile"
                    },
                    "tipo": {
                        "type": "string",
                        "example": "Local"
                    },
                    "color": {
                        "type": "string",
                        "example": "Rojo"
                    },
                    "precio": {
                        "type": "integer",
                        "example": 45000
                    },
                    "precio_final": {
                        "description": "Precio calculado según el descuento aplicado al cliente consultante",
                        "type": "integer",
                        "example": 40500,
                        "nullable": true
                    },
                    "stock": {
                        "type": "integer",
                        "example": 20
                    },
                    "detalles": {
                        "type": "string",
                        "example": "Edición aniversario 2025",
                        "nullable": true
                    },
                    "codigo_producto": {
                        "type": "string",
                        "example": "SCL2025L"
                    },
                    "tallas": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Talla"
                        },
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "Cliente": {
                "title": "Cliente",
                "description": "Modelo completo de cliente B2B",
                "required": [
                    "nombre_comercial",
                    "rut_comercial",
                    "direccion",
                    "categoria",
                    "contacto_nombre",
                    "contacto_email",
                    "porcentaje_oferta"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre_comercial": {
                        "type": "string",
                        "example": "90minutos"
                    },
                    "rut_comercial": {
                        "type": "string",
                        "example": "76123456-7"
                    },
                    "direccion": {
                        "type": "string",
                        "example": "Providencia, Santiago"
                    },
                    "categoria": {
                        "type": "string",
                        "example": "Preferencial",
                        "enum": [
                            "Regular",
                            "Preferencial"
                        ]
                    },
                    "contacto_nombre": {
                        "type": "string",
                        "example": "Juan Pérez"
                    },
                    "contacto_email": {
                        "type": "string",
                        "format": "email",
                        "example": "juan@90minutos.cl"
                    },
                    "porcentaje_oferta": {
                        "type": "integer",
                        "example": 10
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "ClienteInput": {
                "title": "Cliente Input",
                "description": "Datos para crear o actualizar un cliente",
                "required": [
                    "nombre_comercial",
                    "rut_comercial",
                    "direccion",
                    "categoria",
                    "contacto_nombre",
                    "contacto_email",
                    "porcentaje_oferta"
                ],
                "properties": {
                    "nombre_comercial": {
                        "type": "string",
                        "example": "90minutos"
                    },
                    "rut_comercial": {
                        "type": "string",
                        "example": "76123456-7"
                    },
                    "direccion": {
                        "type": "string",
                        "example": "Providencia, Santiago"
                    },
                    "categoria": {
                        "type": "string",
                        "example": "Preferencial",
                        "enum": [
                            "Regular",
                            "Preferencial"
                        ]
                    },
                    "contacto_nombre": {
                        "type": "string",
                        "example": "Juan Pérez"
                    },
                    "contacto_email": {
                        "type": "string",
                        "format": "email",
                        "example": "juan@90minutos.cl"
                    },
                    "porcentaje_oferta": {
                        "type": "integer",
                        "example": 10
                    }
                },
                "type": "object"
            },
...
            "Talla": {
                "title": "Talla",
                "description": "Modelo completo de talla disponible para camisetas",
                "required": [
                    "nombre"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "example": "M"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}